feat(uikit): Add UI Kit page
This commit adds the UI Kit page to the Coodo.xyz website. The UI components and styles to ensure consistency across all Coodo.xyz projects. The page includes information about the tools used Tailwind CSS) and tags (design, react ui, ux). The author of the page is Louis Langanay. An image preview of the UI Kit is also included.